cleaned responsive navigation
