+++ webapp/src/lib/responsive-navigation/responsive-nav-item.svelte
@@ -5,17 +5,20 @@ SPDX-License-Identifier: AGPL-3.0-or-later
 -->
 
 <script lang="ts">
-	import type { ComponentType } from 'svelte';
+	import type { Component } from 'svelte';
 	import type { Icon } from 'lucide-svelte';
 	import Button from '@/components/ui-custom/button.svelte';
 	import IconComponent from '@/components/ui-custom/icon.svelte';
+	import type { IconComponent as ComponentIcon } from '@/components/types';
+	
 
 	interface Props {
 		href: string;
 		label: string;
-		icon?: ComponentType<Icon>;
+		icon?: ComponentIcon;
 		variant?: 'desktop' | 'mobile';
 		onClick?: () => void;
+		onNavigate?: () => void;
 		class?: string;
 	}
 
@@ -25,11 +28,16 @@ SPDX-License-Identifier: AGPL-3.0-or-later
 		icon, 
 		variant = 'desktop', 
 		onClick,
+		onNavigate,
 		class: className = '' 
 	}: Props = $props();
 
 	function handleClick() {
 		onClick?.();
+		// Close mobile menu when navigating (only for mobile variant)
+		if (variant === 'mobile') {
+			onNavigate?.();
+		}
 	}
 </script>
